Reimplement post processing callbacks.
Callbacks can be added using '{before,after}[_<attachment>]_post_process'
class methods (similar to Paperclip), using either a Symbol to use an
instance method as callback, or a proc, which will be called with the
resource instance as first argument. Callbacks are run in-order ('before'
callbacks) and in reverse ('after' callbacks).

The 'before' and 'after' callback chain can be halted by returning false
from the callback. Further callbacks are not run. Halting the 'before'
callback chain skips post processing and all 'after' callbacks.